refactor using styled component
