+++ src/components/DateTimeRangePicker/style.ts
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const StartEndDates = styled.div`
+    display: flex;
+    justify-content: space-between;
+`;
